java 4장 Programming 문제 코드

+++ ch04/src/ch04programming/Programming06.java
@@ -6,8 +6,8 @@ public class Programming06 {
 		
 		
 		for (int i = -60; i <=140; i=i+20) {
-			int f = (9/5)*i + 32;
-			System.out.printf("섭씨온도(C) %d의 화씨온도(F)는 : %d%n", i, f);
+			double f = (9.0/5.0)*i + 32;
+			System.out.printf("섭씨온도(C) %d의 화씨온도(F)는 : %.0f%n", i, f);
 		}
